+++ _pytest/test_train.py
@@ -60,7 +60,8 @@ def test_train_spacy_sklearn_finetune_ner(component_builder):
     entities = result['entities']
     # Although the model is trained on restaurant entities, we can use the entities (`GPE`, `DATE`)
     # from spacy since we are fine tuning. This should even be the case if the rasa-entity training data changes!
-    assert {u'start': 15, u'end': 28, u'value': u'New York City', u'entity': u'GPE'} in entities
+    assert {u'start': 15, u'end': 28, u'value': u'New York City',
+            'extractor': 'ner_spacy', u'entity': u'GPE'} in entities
 
 
 def test_train_named_model(component_builder):