+++ rasa_nlu/components.py
@@ -139,8 +139,8 @@ def validate_arguments(pipeline, config, allow_empty_pipeline=False):
             raise Exception("Failed to validate at component '{}'. {}".format(component.name, e))
 
     # Reset context to test processing phase and prepare for training phase
-    context = after_init_context
-    context["text"] = None
+    context = {"entities": [], "text": None}
+    context.update(after_init_context)
 
     for component in pipeline:
         try: