+++ rasa_nlu/config.py
@@ -35,7 +35,7 @@ DEFAULT_CONFIG = {
     "max_number_of_ngrams": 7,
     "pipeline": [],
     "response_log": "logs",
-    "duckling_processing_mode": "append",
+    "duckling_dimensions": None,
     "luis_data_tokenizer": None,
     "entity_crf_BILOU_flag": True,
     "entity_crf_features": [
@@ -114,19 +114,30 @@ class RasaNLUConfig(object):
     def view(self):
         return json.dumps(self.__dict__, indent=4)
 
+    def split_arg(self, config, arg_name):
+        if arg_name in config and isinstance(config[arg_name], six.string_types):
+            config[arg_name] = config[arg_name].split(",")
+        return config
+
     def split_pipeline(self, config):
-        if "pipeline" in config.keys() and "," in config["pipeline"]:
-            config["pipeline"] = config["pipeline"].split(",")
+        if "pipeline" in config and isinstance(config["pipeline"], six.string_types):
+            config = self.split_arg(config, "pipeline")
+            if "pipeline" in config and len(config["pipeline"]) == 1:
+                config["pipeline"] = config["pipeline"][0]
         return config
 
     def create_cmdline_config(self, cmdline_args):
         cmdline_config = {k: v for k, v in list(cmdline_args.items()) if v is not None}
-        return self.split_pipeline(cmdline_config)
+        cmdline_config = self.split_pipeline(cmdline_config)
+        cmdline_config = self.split_arg(cmdline_config, "duckling_dimensions")
+        return cmdline_config
 
     def create_env_config(self, env_vars):
         keys = [key for key in env_vars.keys() if "RASA" in key]
         env_config = {key.split('RASA_')[1].lower(): env_vars[key] for key in keys}
-        return self.split_pipeline(env_config)
+        env_config = self.split_pipeline(env_config)
+        env_config = self.split_arg(env_config, "duckling_dimensions")
+        return env_config
 
     def is_set(self, key):
         return key in self.__dict__ and self[key] is not None