+++ rasa_nlu/extractors/duckling_extractor.py
@@ -3,6 +3,9 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+import os
+import io
+import json
 import typing
 from typing import Any
 from typing import Dict
@@ -10,18 +13,15 @@ from typing import List
 from typing import Optional
 from typing import Text
 
-from rasa_nlu.components import Component
+from rasa_nlu.extractors import EntityExtractor
 from rasa_nlu.model import Metadata
-
+from inspect import getmembers
 
 if typing.TYPE_CHECKING:
     from duckling import DucklingWrapper
 
 
-DUCKLING_PROCESSING_MODES = ["replace", "append"]
-
-
-class DucklingExtractor(Component):
+class DucklingExtractor(EntityExtractor):
     """Adds entity normalization by analyzing found entities and transforming them into regular formats."""
 
     name = "ner_duckling"
@@ -32,10 +32,15 @@ class DucklingExtractor(Component):
 
     output_provides = ["entities"]
 
-    def __init__(self, duckling_processing_mode, duckling=None):
+    @staticmethod
+    def available_dimensions():
+        from duckling.dim import Dim
+        return [m[1] for m in getmembers(Dim) if not m[0].startswith("__") and not m[0].endswith("__")]
+
+    def __init__(self, dimensions=None, duckling=None):
         # type: (Text, Optional[DucklingWrapper]) -> None
 
-        self.duckling_processing_mode = duckling_processing_mode
+        self.dimensions = dimensions if dimensions else self.available_dimensions()
         self.duckling = duckling
 
     @classmethod
@@ -44,12 +49,15 @@ class DucklingExtractor(Component):
         return ["duckling"]
 
     @classmethod
-    def create(cls, duckling_processing_mode):
-        if duckling_processing_mode not in DUCKLING_PROCESSING_MODES:
-            raise ValueError("Invalid duckling processing mode. Got '{}'. Allowed: {}".format(
-                duckling_processing_mode, ", ".join(DUCKLING_PROCESSING_MODES)))
+    def create(cls, duckling_dimensions):
+        if duckling_dimensions is None:
+            duckling_dimensions = cls.available_dimensions()
+        unknown_dimensions = [dim for dim in duckling_dimensions if dim not in cls.available_dimensions()]
+        if len(unknown_dimensions) > 0:
+            raise ValueError("Invalid duckling dimension. Got '{}'. Allowed: {}".format(
+                ", ".join(unknown_dimensions), ", ".join(cls.available_dimensions())))
 
-        return DucklingExtractor(duckling_processing_mode)
+        return DucklingExtractor(duckling_dimensions)
 
     @classmethod
     def cache_key(cls, model_metadata):
@@ -68,34 +76,39 @@ class DucklingExtractor(Component):
                 raise Exception("Duckling error. {}".format(e))
 
     def process(self, text, entities):
-        # type: (Text, List[Dict[Text, Any]], Text) -> Dict[Text, Any]
+        # type: (Text, List[Dict[Text, Any]]) -> Dict[Text, Any]
 
+        extracted = []
         if self.duckling is not None:
-            parsed = self.duckling.parse(text)
-            for duckling_match in parsed:
-                for entity in entities:
-                    if entity["start"] == duckling_match["start"] and entity["end"] == duckling_match["end"]:
-                        entity["value"] = duckling_match["value"]["value"]
-                        entity["duckling"] = duckling_match["dim"]
-                        break
-                else:
-                    if self.duckling_processing_mode == "append":
-                        # Duckling will retrieve multiple entities, even if they overlap..
-                        # hence the append mode might add some noise to the found entities
-                        entities.append({
-                            "entity": duckling_match["dim"],
-                            "duckling": duckling_match["dim"],
-                            "value": duckling_match["value"]["value"],
-                            "start": duckling_match["start"],
-                            "end": duckling_match["end"],
-                        })
+            matches = self.duckling.parse(text)
+            relevant_matches = [match for match in matches if match["dim"] in self.dimensions]
+            for match in relevant_matches:
+                entity = {"start": match["start"],
+                          "end": match["end"],
+                          "value": match["value"]["value"],
+                          "entity": match["dim"]}
 
+                extracted.append(entity)
+
+        extracted = self.add_extractor_name(extracted)
+        entities.extend(extracted)
         return {
             "entities": entities
         }
 
-    @classmethod
-    def load(cls, duckling_processing_mode):
-        # type: (Text) -> DucklingExtractor
+    def persist(self, model_dir):
+        # type: (Text) -> Dict[Text, Any]
+        file_name = self.name+".json"
+        full_name = os.path.join(model_dir, file_name)
+        with io.open(full_name, 'w') as f:
+            f.write(json.dumps({"dimensions": self.dimensions}))
+        return {"ner_duckling_persisted": file_name}
 
-        return cls.create(duckling_processing_mode)
+    @classmethod
+    def load(cls, model_dir, ner_duckling_persisted):
+        # type: (Text) -> DucklingExtractor
+        persisted = os.path.join(model_dir, ner_duckling_persisted)
+        if os.path.isfile(persisted):
+            with io.open(persisted, encoding='utf-8') as f:
+                persisted_data = json.loads(f.read())
+                return cls.create(persisted_data["dimensions"])